feat: add unit tests for storage module
