+++ x/storage/types/message_test.go
@@ -7,7 +7,6 @@ import (
 	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/prysmaticlabs/prysm/crypto/bls"
 	"github.com/stretchr/testify/require"
 
 	"github.com/bnb-chain/greenfield/testutil/sample"
@@ -145,261 +144,6 @@ func TestMsgUpdateBucketInfo_ValidateBasic(t *testing.T) {
 				PaymentAddress:   sample.AccAddress(),
 				ChargedReadQuota: &common.UInt64Value{Value: 10000},
 			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := tt.msg.ValidateBasic()
-			if tt.err != nil {
-				require.ErrorIs(t, err, tt.err)
-				return
-			}
-			require.NoError(t, err)
-		})
-	}
-}
-
-func TestMsgCreateObject_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name string
-		msg  MsgCreateObject
-		err  error
-	}{
-		{
-			name: "normal",
-			msg: MsgCreateObject{
-				Creator:           sample.AccAddress(),
-				BucketName:        testBucketName,
-				ObjectName:        testObjectName,
-				PayloadSize:       1024,
-				Visibility:        VISIBILITY_TYPE_PRIVATE,
-				ContentType:       "content-type",
-				PrimarySpApproval: &common.Approval{},
-				ExpectChecksums:   [][]byte{sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum()},
-			},
-		}, {
-			name: "invalid object name",
-			msg: MsgCreateObject{
-				Creator:           sample.AccAddress(),
-				BucketName:        testBucketName,
-				ObjectName:        "",
-				PayloadSize:       1024,
-				Visibility:        VISIBILITY_TYPE_PRIVATE,
-				ContentType:       "content-type",
-				PrimarySpApproval: &common.Approval{},
-				ExpectChecksums:   [][]byte{sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum()},
-			},
-			err: gnfderrors.ErrInvalidObjectName,
-		}, {
-			name: "invalid object name",
-			msg: MsgCreateObject{
-				Creator:           sample.AccAddress(),
-				BucketName:        testBucketName,
-				ObjectName:        "../object",
-				PayloadSize:       1024,
-				Visibility:        VISIBILITY_TYPE_PRIVATE,
-				ContentType:       "content-type",
-				PrimarySpApproval: &common.Approval{},
-				ExpectChecksums:   [][]byte{sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum()},
-			},
-			err: gnfderrors.ErrInvalidObjectName,
-		}, {
-			name: "invalid object name",
-			msg: MsgCreateObject{
-				Creator:           sample.AccAddress(),
-				BucketName:        testBucketName,
-				ObjectName:        "//object",
-				PayloadSize:       1024,
-				Visibility:        VISIBILITY_TYPE_PRIVATE,
-				ContentType:       "content-type",
-				PrimarySpApproval: &common.Approval{},
-				ExpectChecksums:   [][]byte{sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum()},
-			},
-			err: gnfderrors.ErrInvalidObjectName,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := tt.msg.ValidateBasic()
-			if tt.err != nil {
-				require.ErrorIs(t, err, tt.err)
-				return
-			}
-			require.NoError(t, err)
-		})
-	}
-}
-
-func TestMsgCancelCreateObject_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name string
-		msg  MsgCancelCreateObject
-		err  error
-	}{
-		{
-			name: "basic",
-			msg: MsgCancelCreateObject{
-				Operator:   sample.AccAddress(),
-				BucketName: testBucketName,
-				ObjectName: testObjectName,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := tt.msg.ValidateBasic()
-			if tt.err != nil {
-				require.ErrorIs(t, err, tt.err)
-				return
-			}
-			require.NoError(t, err)
-		})
-	}
-}
-
-func TestMsgDeleteObject_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name string
-		msg  MsgDeleteObject
-		err  error
-	}{
-		{
-			name: "normal",
-			msg: MsgDeleteObject{
-				Operator:   sample.AccAddress(),
-				BucketName: testBucketName,
-				ObjectName: testObjectName,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := tt.msg.ValidateBasic()
-			if tt.err != nil {
-				require.ErrorIs(t, err, tt.err)
-				return
-			}
-			require.NoError(t, err)
-		})
-	}
-}
-
-func TestMsgCopyObject_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name string
-		msg  MsgCopyObject
-		err  error
-	}{
-		{
-			name: "valid address",
-			msg: MsgCopyObject{
-				Operator:      sample.AccAddress(),
-				SrcBucketName: testBucketName,
-				SrcObjectName: testObjectName,
-				DstBucketName: "dst" + testBucketName,
-				DstObjectName: "dst" + testObjectName,
-				DstPrimarySpApproval: &common.Approval{
-					ExpiredHeight: 100,
-					Sig:           []byte("xxx"),
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := tt.msg.ValidateBasic()
-			if tt.err != nil {
-				require.ErrorIs(t, err, tt.err)
-				return
-			}
-			require.NoError(t, err)
-		})
-	}
-}
-
-func TestMsgSealObject_ValidateBasic(t *testing.T) {
-	checksums := [][]byte{sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum(), sample.Checksum()}
-	blsSignDoc := NewSecondarySpSealObjectSignDoc("greenfield_9000-1", 1, math.NewUint(1), GenerateHash(checksums[:])).GetSignBytes()
-	blsPrivKey, _ := bls.RandKey()
-	aggSig := blsPrivKey.Sign(blsSignDoc[:]).Marshal()
-	tests := []struct {
-		name string
-		msg  MsgSealObject
-		err  error
-	}{
-		{
-			name: "normal",
-			msg: MsgSealObject{
-				Operator:                    sample.AccAddress(),
-				BucketName:                  testBucketName,
-				ObjectName:                  testObjectName,
-				SecondarySpBlsAggSignatures: aggSig,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := tt.msg.ValidateBasic()
-			if tt.err != nil {
-				require.ErrorIs(t, err, tt.err)
-				return
-			}
-			require.NoError(t, err)
-		})
-	}
-}
-
-func TestMsgRejectSealObject_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name string
-		msg  MsgRejectSealObject
-		err  error
-	}{
-		{
-			name: "normal",
-			msg: MsgRejectSealObject{
-				Operator:   sample.AccAddress(),
-				BucketName: testBucketName,
-				ObjectName: testObjectName,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := tt.msg.ValidateBasic()
-			if tt.err != nil {
-				require.ErrorIs(t, err, tt.err)
-				return
-			}
-			require.NoError(t, err)
-		})
-	}
-}
-
-func TestMsgUpdateObjectInfo_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name string
-		msg  MsgUpdateObjectInfo
-		err  error
-	}{
-		{
-			name: "normal",
-			msg: MsgUpdateObjectInfo{
-				Operator:   sample.AccAddress(),
-				BucketName: testBucketName,
-				ObjectName: testObjectName,
-				Visibility: VISIBILITY_TYPE_INHERIT,
-			},
-		},
-		{
-			name: "abnormal",
-			msg: MsgUpdateObjectInfo{
-				Operator:   sample.AccAddress(),
-				BucketName: testBucketName,
-				ObjectName: testObjectName,
-				Visibility: VISIBILITY_TYPE_UNSPECIFIED,
-			},
-			err: ErrInvalidVisibility,
 		},
 	}
 	for _, tt := range tests {
@@ -633,40 +377,6 @@ func TestMsgMirrorBucket_ValidateBasic(t *testing.T) {
 		}, {
 			name: "invalid account name",
 			msg: MsgMirrorBucket{
-				Operator: "wrong address",
-				Id:       math.NewUint(1),
-			},
-			err: sdkerrors.ErrInvalidAddress,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := tt.msg.ValidateBasic()
-			if tt.err != nil {
-				require.ErrorIs(t, err, tt.err)
-				return
-			}
-			require.NoError(t, err)
-		})
-	}
-}
-
-func TestMsgMirrorObject_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name string
-		msg  MsgMirrorObject
-		err  error
-	}{
-		{
-			name: "normal",
-			msg: MsgMirrorObject{
-				Operator: sample.AccAddress(),
-				Id:       math.NewUint(1),
-			},
-		},
-		{
-			name: "invalid address",
-			msg: MsgMirrorObject{
 				Operator: "wrong address",
 				Id:       math.NewUint(1),
 			},
